third_party/WebKit/LayoutTests/inspector/elements/styles-4/undo-add-property.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>

<style>
.container {
  font-weight: bold
}
</style>

<script>

function test()
{
    ElementsTestRunner.selectNodeAndWaitForStyles("container", testAppendProperty);

    function testAppendProperty()
    {
        TestRunner.addResult("=== Last property ===");
        testAddProperty("margin-left: 2px", undefined, testInsertBegin);
    }

    function testInsertBegin()
    {
        TestRunner.addResult("=== First property ===");
        testAddProperty("margin-top: 0px", 0, testInsertMiddle);
    }

    function testInsertMiddle()
    {
        TestRunner.addResult("=== Middle property ===");
        testAddProperty("margin-right: 1px", 1, TestRunner.completeTest.bind(InspectorTest));
    }

    function testAddProperty(propertyText, index, callback)
    {
        TestRunner.addResult("(Initial value)");
        ElementsTestRunner.dumpSelectedElementStyles(true);

        var treeItem = ElementsTestRunner.getMatchedStylePropertyTreeItem("font-weight");
        var treeElement = treeItem.section().addNewBlankProperty(index);
        treeElement.applyStyleText(propertyText, true);
        ElementsTestRunner.waitForStyles("container", step1);

        function step1()
        {
            TestRunner.addResult("(After adding property)");
            ElementsTestRunner.dumpSelectedElementStyles(true);

            TestRunner.domModel.undo();
            ElementsTestRunner.selectNodeAndWaitForStyles("other", step2);
        }

        function step2()
        {
            TestRunner.addResult("(After undo)");
            ElementsTestRunner.dumpSelectedElementStyles(true);

            TestRunner.domModel.redo();
            ElementsTestRunner.selectNodeAndWaitForStyles("container", step3);
        }

        function step3()
        {
            TestRunner.addResult("(After redo)");
            ElementsTestRunner.dumpSelectedElementStyles(true);
            callback();
        }
    }
}
</script>
</head>

<body onload="runTest()">
<p>
Tests that adding a property is undone properly.
</p>

<div id="container" class="container"></div>
<div id="other" class="container"></div>

</body>
</html>
