third_party/WebKit/LayoutTests/inspector/elements/styles-4/undo-add-property-expected.txt

Tests that adding a property is undone properly.

=== Last property ===
(Initial value)
[expanded] 
element.style { ()

[expanded] 
.container { (undo-add-property.html:7 -> undo-add-property.html:7:13)
    font-weight: bold;

[expanded] 
div { (user agent stylesheet)
    display: block;

(After adding property)
[expanded] 
element.style { ()

[expanded] 
.container { (undo-add-property.html:7 -> undo-add-property.html:7:13)
    font-weight: bold;
    margin-left: 2px;

[expanded] 
div { (user agent stylesheet)
    display: block;

(After undo)
[expanded] 
element.style { ()

[expanded] 
.container { (undo-add-property.html:7 -> undo-add-property.html:7:13)
    font-weight: bold;

[expanded] 
div { (user agent stylesheet)
    display: block;

(After redo)
[expanded] 
element.style { ()

[expanded] 
.container { (undo-add-property.html:7 -> undo-add-property.html:7:13)
    font-weight: bold;
    margin-left: 2px;

[expanded] 
div { (user agent stylesheet)
    display: block;

=== First property ===
(Initial value)
[expanded] 
element.style { ()

[expanded] 
.container { (undo-add-property.html:7 -> undo-add-property.html:7:13)
    font-weight: bold;
    margin-left: 2px;

[expanded] 
div { (user agent stylesheet)
    display: block;

(After adding property)
[expanded] 
element.style { ()

[expanded] 
.container { (undo-add-property.html:7 -> undo-add-property.html:7:13)
    margin-top: 0px;
    font-weight: bold;
    margin-left: 2px;

[expanded] 
div { (user agent stylesheet)
    display: block;

(After undo)
[expanded] 
element.style { ()

[expanded] 
.container { (undo-add-property.html:7 -> undo-add-property.html:7:13)
    font-weight: bold;
    margin-left: 2px;

[expanded] 
div { (user agent stylesheet)
    display: block;

(After redo)
[expanded] 
element.style { ()

[expanded] 
.container { (undo-add-property.html:7 -> undo-add-property.html:7:13)
    margin-top: 0px;
    font-weight: bold;
    margin-left: 2px;

[expanded] 
div { (user agent stylesheet)
    display: block;

=== Middle property ===
(Initial value)
[expanded] 
element.style { ()

[expanded] 
.container { (undo-add-property.html:7 -> undo-add-property.html:7:13)
    margin-top: 0px;
    font-weight: bold;
    margin-left: 2px;

[expanded] 
div { (user agent stylesheet)
    display: block;

(After adding property)
[expanded] 
element.style { ()

[expanded] 
.container { (undo-add-property.html:7 -> undo-add-property.html:7:13)
    margin-top: 0px;
    margin-right: 1px;
    font-weight: bold;
    margin-left: 2px;

[expanded] 
div { (user agent stylesheet)
    display: block;

(After undo)
[expanded] 
element.style { ()

[expanded] 
.container { (undo-add-property.html:7 -> undo-add-property.html:7:13)
    margin-top: 0px;
    font-weight: bold;
    margin-left: 2px;

[expanded] 
div { (user agent stylesheet)
    display: block;

(After redo)
[expanded] 
element.style { ()

[expanded] 
.container { (undo-add-property.html:7 -> undo-add-property.html:7:13)
    margin-top: 0px;
    margin-right: 1px;
    font-weight: bold;
    margin-left: 2px;

[expanded] 
div { (user agent stylesheet)
    display: block;
